+++ ndonnx/_core/_stringimpl.py
@@ -11,7 +11,7 @@ import ndonnx as ndx
 import ndonnx._data_types as dtypes
 import ndonnx._opset_extensions as opx
 
-from ._interface import OperationsBlock
+from ._default import OperationsBlock
 from ._utils import binary_op
 
 if TYPE_CHECKING:
